pyproject.toml
[tool.poetry]
name = "dib"
version = "0.1.2a0"
description = "A Discord Bot connected to ML RAG"
authors = ["Joaquin Dominguez <joaquin.dominguez@proton.me>"]
readme = "README.md"
homepage = "https://fiddles.app"
documentation = "https://fiddles.app"
license = "MIT"
package-mode = true
packages = [{include = "dib", from = "src"}]
repository = "https://gitlab.com/j-dominguez9/discord_bot"
keywords = ["discord", "bot", "RAG", "LLM"]

[tool.poetry.urls]
"Bug Tracker" = "https://gitlab.com/j-dominguez9/discord_bot/-/issues"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.3.2"
python-dotenv = "^1.0.1"
groq = "0.5.0"
pypdf2 = "^3.0.1"
tqdm = "^4.66.2"
icecream = "^2.1.3"
requests = "^2.31.0"
openai = "^1.14.1"
llama-index-core = "0.10.30"
qdrant-client = "^1.8.0"
llama-index-embeddings-cohere = "^0.1.4"
llama-index-vector-stores-qdrant = "^0.1.4"
llama-index-llms-anyscale = "^0.1.3"
llama-index-llms-groq = "^0.1.3"
llama-index-postprocessor-cohere-rerank = "^0.1.2"
dvc = {extras = ["gs"], version = "^3.48.4"}
llama-index-vector-stores-milvus = "^0.1.6"
pymilvus = "^2.4.0"
scipy = "^1.12.0"


[tool.poetry.group.test.dependencies]
pyright = "^1.1.354"
pytest = "^8.1.1"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
pyright = "^1.1.355"
flake8 = "^7.0.0"
isort = "^5.13.2"
pre-commit = "^3.6.2"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-dracula-theme = "^1.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
